refactor: rename to `terminal`
