+++ kernel/src/terminal/log.rs
