+++ kernel/src/terminal/vram.rs
