+++ kernel/src/terminal/writer.rs
